Bug fix in ET calculation edge case for Soil Mods
The purpose of the editted code block is to move the value stored in
`soil_rechr` into `et` and zero `soil_rechr`.

Physically, this means that all the moisutre in the recharge layer in
the soil is evapotranspirated. The order of operations is reversed and
`soil_rechr` is set to zero before moving its value into `et`.